wnf

Struct OwnedState

pub struct OwnedState<T>
where
    T: ?Sized,
{ /* private fields */ }

An owned state

This deletes the represented state on drop. You can prevent this behavior by calling the leak method.

While ownership in Rust usually refers to the ownership of memory, this applies the idea of ownership to an external entity, namely a state. It’s similar to OwnedHandle in that regard.

An OwnedState<T> can be borrowed as a BorrowedState<'_, T> using the AsState::as_state method.

The type parameter T is the type of data contained in the state.

Implementations

impl<T> OwnedState<T>
where T: Read<T> + NoUninit,

pub fn apply<D, F>(&self, transform: F) -> Result<D>
where D: Borrow<T>, F: FnMut(T) -> D,

Applies a transformation to the data of this state

This essentially queries the state data, applies the given transformation closure and then updates the state data with the transformed value. However, it tries to do so in a loop using change stamps to ensure that no concurrent update happens between querying and updating the state data. This means that the given closure may be called multiple times. Note that it does not reliably avoid concurrent updates while the actual update is happening. If another concurrent update makes the size of the state data exceed the internal capacity of the state (causing a reallocation), it may happen that this update does not have the desired effect on the state data.

The closure receives an owned T on the stack, requiring T: Sized. In order to receive a Box<T> for T: ?Sized, use the apply_boxed method.

For fallible transformations, i.e. when the transformation closure returns a Result<D, E>, use the try_apply method.

The return value is the value with which the state was ultimately updated, i.e. the return value of the last call to the given closure.

For example, to increment the value of a state by one and return the incremented value:

use std::io;

use wnf::{AsState, OwnedState};

fn increment<S>(state: S) -> io::Result<u32>
where
    S: AsState<Data = u32>,
{
    state.as_state().apply(|value| value + 1)
}

let state = OwnedState::create_temporary()?;
state.set(&42)?;

let new_data = increment(&state)?;
assert_eq!(new_data, 43);

Errors

Returns an error if querying or updating fails

pub fn try_apply<D, E, F>(&self, transform: F) -> Result<D>
where D: Borrow<T>, E: Into<Box<dyn Error + Send + Sync>>, F: FnMut(T) -> Result<D, E>,

Applies a fallible transformation to the data of this state

This essentially queries the state data, applies the given transformation closure and then (in case it succeeds) updates the state data with the transformed value. However, it tries to do so in a loop using change stamps to ensure that no concurrent update happens between querying and updating the state data. This means that the given closure may be called multiple times. Note that it does not reliably avoid concurrent updates while the actual update is happening. If another concurrent update makes the size of the state data exceed the internal capacity of the state (causing a reallocation), it may happen that this update does not have the desired effect on the state data.

The closure receives an owned T on the stack, requiring T: Sized. In order to receive a Box<T> for T: ?Sized, use the apply_boxed method.

For infallible transformations, i.e. when the transformation closure returns a D rather than a Result<D, E>, use the apply method.

The return value is the value with which the state was ultimately updated, i.e. the return value of the last call to the given closure.

For example, to increment the value of a state by one, unless a maximum is reached, and return the incremented value:

use std::io;

use wnf::{AsState, OwnedState};

fn try_increment<S>(state: S, max: u32) -> io::Result<u32>
where
    S: AsState<Data = u32>,
{
    state.as_state().try_apply(|value| {
        if value < max {
            Ok(value + 1)
        } else {
            Err("maximum reached")
        }
    })
}

let state = OwnedState::create_temporary()?;
state.set(&42)?;

let new_data = try_increment(&state, 43)?;
assert_eq!(new_data, 43);

let result = try_increment(&state, 43);
assert!(result.is_err());

Errors

Returns an error if querying or updating fails, or if the transformation closure returns an error

impl<T> OwnedState<T>
where T: Read<Box<T>> + NoUninit + ?Sized,

pub fn apply_boxed<D, F>(&self, transform: F) -> Result<D>
where D: Borrow<T>, F: FnMut(Box<T>) -> D,

Applies a transformation to the data of this state as a box

This essentially queries the state data, applies the given transformation closure and then updates the state data with the transformed value. However, it tries to do so in a loop using change stamps to ensure that no concurrent update happens between querying and updating the state data. This means that the given closure may be called multiple times. Note that it does not reliably avoid concurrent updates while the actual update is happening. If another concurrent update makes the size of the state data exceed the internal capacity of the state (causing a reallocation), it may happen that this update does not have the desired effect on the state data.

The closure receives a Box<T>. In order to receive an owned T on the stack (requiring T: Sized), use the apply method.

For fallible transformations, i.e. when the transformation closure returns a Result<D, E>, use the try_apply_boxed method.

The return value is the value with which the state was ultimately updated, i.e. the return value of the last call to the given closure.

For example, to extend a slice by one element and return the extended slice as a Vec<u32>:

use std::io;

use wnf::{AsState, OwnedState};

fn extend<S>(state: S) -> io::Result<Vec<u32>>
where
    S: AsState<Data = [u32]>,
{
    state.as_state().apply_boxed(|slice| {
        let mut vec = slice.into_vec();
        vec.push(42);
        vec
    })
}

let state = OwnedState::<[u32]>::create_temporary()?;
state.set(&[1, 2, 3])?;

let new_data = extend(&state)?;
assert_eq!(*new_data, [1, 2, 3, 42]);

Errors

Returns an error if querying or updating fails

pub fn try_apply_boxed<D, E, F>(&self, transform: F) -> Result<D>
where D: Borrow<T>, E: Into<Box<dyn Error + Send + Sync>>, F: FnMut(Box<T>) -> Result<D, E>,

Applies a fallible transformation to the data of this state as a box

This essentially queries the state data, applies the given transformation closure and then (in case it succeeds) updates the state data with the transformed value. However, it tries to do so in a loop using change stamps to ensure that no concurrent update happens between querying and updating the state data. This means that the given closure may be called multiple times. Note that it does not reliably avoid concurrent updates while the actual update is happening. If another concurrent update makes the size of the state data exceed the internal capacity of the state (causing a reallocation), it may happen that this update does not have the desired effect on the state data.

The closure receives a Box<T>. In order to receive an owned T on the stack (requiring T: Sized), use the try_apply method.

For infallible transformations, i.e. when the transformation closure returns a D rather than a Result<D, E>, use the apply_boxed method.

The return value is the value with which the state was ultimately updated, i.e. the return value of the last call to the given closure.

For example, to extend a slice by one element, unless a maximum length is reached, and return the extended slice as a Vec<u32>:

use std::io;

use wnf::{AsState, OwnedState};

fn try_extend<S>(state: S, max_len: usize) -> io::Result<Vec<u32>>
where
    S: AsState<Data = [u32]>,
{
    state.as_state().try_apply_boxed(|slice| {
        if slice.len() < max_len {
            let mut vec = slice.into_vec();
            vec.push(42);
            Ok(vec)
        } else {
            Err("maximum length reached")
        }
    })
}

let state = OwnedState::<[u32]>::create_temporary()?;
state.set(&[1, 2, 3])?;

let new_data = try_extend(&state, 4)?;
assert_eq!(*new_data, [1, 2, 3, 42]);

let result = try_extend(&state, 4);
assert!(result.is_err());

Errors

Returns an error if querying or updating fails

impl<T> OwnedState<T>
where T: ?Sized,

pub fn exists(&self) -> Result<bool>

Returns whether this state exists

Errors

Returns an error if obtaining the information fails

pub fn subscribers_present(&self) -> Result<bool>

Returns whether this state has at least one subscriber

Errors

Returns an error if obtaining the information fails

pub fn is_quiescent(&self) -> Result<bool>

Returns whether this state is “quiescent”, i.e. none of the listeners subscribed to it are currently running

Errors

Returns an error if obtaining the information fails

impl<T> OwnedState<T>
where T: ?Sized,

pub fn create_temporary() -> Result<Self>

Creates an OwnedState<T> with temporary lifetime and machine scope

This is a convenience method for quickly creating a state, e.g. for testing purposes. For more precise control over the created state, use the StateCreation builder.

Note that a newly created state is initialized with data of size zero. This means that unless the data type T is zero-sized or a slice type, you need to update the state data with a value of type T before querying it for the first time.

Errors

Returns an error if creating the state fails

pub fn delete(self) -> Result<()>

Deletes this state

Note that an OwnedState<T> will be deleted automatically when it is dropped, so calling this method is usually not necessary. It is useful, however, if you want to handle errors.

Errors

Returns an error if deleting the state fails

impl<T> OwnedState<T>
where T: Read<T>,

pub fn get(&self) -> Result<T>

Queries the data of this state

This produces an owned T on the stack and hence requires T: Sized. In order to produce a Box<T> for T: ?Sized, use the get_boxed method.

This returns the data of the state without a change stamp. In order to query both the data and the change stamp, use the query method.

Errors

Returns an error if querying fails, including the case that the queried data is not a valid T

pub fn query(&self) -> Result<StampedData<T>>

Queries the data of this state together with its change stamp

This produces an owned T on the stack and hence requires T: Sized. In order to produce a Box<T> for T: ?Sized, use the query_boxed method.

This returns the data of the state together with its change stamp as a StampedData<T>. In order to only query the data, use the get method.

Errors

Returns an error if querying fails, including the case that the queried data is not a valid T

impl<T> OwnedState<T>
where T: Read<Box<T>> + ?Sized,

pub fn get_boxed(&self) -> Result<Box<T>>

Queries the data of this state as a box

This produces a Box<T>. In order to produce an owned T on the stack (requiring T: Sized), use the get method.

This returns the data of the state without a change stamp. In order to query both the data and the change stamp, use the query_boxed method.

Errors

Returns an error if querying fails, including the case that the queried data is not a valid T

pub fn query_boxed(&self) -> Result<StampedData<Box<T>>>

Queries the data of this state as a box together with its change stamp

This produces a Box<T>. In order to produce an owned T on the stack (requiring T: Sized), use the query method.

This returns the data of the state together with its change stamp as a StampedData<Box<T>>. In order to only query the data, use the get_boxed method.

Errors

Returns an error if querying fails, including the case that the queried data is not a valid T

impl<T> OwnedState<T>
where T: ?Sized,

pub fn change_stamp(&self) -> Result<ChangeStamp>

Queries the change stamp of this state

Errors

Returns an error if querying the change stamp fails

impl<T> OwnedState<T>
where T: Read<T> + NoUninit,

pub fn replace(&self, new_value: &T) -> Result<T>

Replaces the data of this state, returning the previous value

This essentially queries the state data, updates it with the given value and returns the previous value. However, it tries to do so in a loop using change stamps to ensure that no concurrent update happens between querying and updating the data. Note that it does not reliably avoid concurrent updates while the actual update is happening. If another concurrent update makes the size of the state data exceed the internal capacity of the state (causing a reallocation), it may happen that this update does not have the desired effect on the state data.

This produces an owned T on the stack and hence requires T: Sized. In order to produce a Box<T> for T: ?Sized, use the replace_boxed method.

For example, to set the value of a state to zero while returning the previous value:

use std::io;

use wnf::{AsState, OwnedState};

fn reset<S>(state: S) -> io::Result<u32>
where
    S: AsState<Data = u32>,
{
    state.as_state().replace(&0)
}

let state = OwnedState::create_temporary()?;
state.set(&42)?;

let previous_value = reset(&state)?;
assert_eq!(previous_value, 42);

Errors

Returns an error if querying or updating fails

impl<T> OwnedState<T>
where T: Read<Box<T>> + NoUninit + ?Sized,

pub fn replace_boxed(&self, new_value: &T) -> Result<Box<T>>

Replaces the data of this state, returning the previous value as a box

This essentially queries the state data, updates it with the given value and returns the previous value. However, it tries to do so in a loop using change stamps to ensure that no concurrent update happens between querying and updating the data. Note that it does not reliably avoid concurrent updates while the actual update is happening. If another concurrent update makes the size of the state data exceed the internal capacity of the state (causing a reallocation), it may happen that this update does not have the desired effect on the state data.

This produces a Box<T>. In order to produce an owned T on the stack (requiring T: Sized), use the replace method.

For example, to make a slice empty while returning the previous (boxed) slice:

use std::io;

use wnf::{AsState, OwnedState};

fn clear<S>(state: S) -> io::Result<Box<[u32]>>
where
    S: AsState<Data = [u32]>,
{
    state.as_state().replace_boxed(&[])
}

let state = OwnedState::<[u32]>::create_temporary()?;
state.set(&[1, 2, 3])?;

let previous_value = clear(&state)?;
assert_eq!(*previous_value, [1, 2, 3]);

Errors

Returns an error if querying or updating fails

impl<T> OwnedState<T>
where T: ?Sized,

pub const fn state_name(&self) -> StateName

Returns the name of this state

pub fn leak(self) -> BorrowedState<'static, T>

Leaks this OwnedState<T>

This consumes the OwnedState<T> without dropping it, returning a BorrowedState<'static, T> with 'static lifetime that represents the same underlying state.

Note that while this is named after Box::leak, it doesn’t leak memory, but it leaks a state in the sense that the state doesn’t get deleted on drop.

pub fn cast<U>(self) -> OwnedState<U>
where U: ?Sized,

Casts the data type of this state to a different type U

The returned OwnedState<U> represents the same underlying state, but treats it as containing data of a different type U.

impl<T> OwnedState<T>
where T: NoUninit + ?Sized,

pub fn set(&self, data: &T) -> Result<()>

Updates the data of this state with the given value

The update is performed regardless of the current change stamp of the state. In order to perform the update conditionally based on the change stamp, use the update method.

Errors

Returns an error if updating fails

pub fn update( &self, data: &T, expected_change_stamp: impl Into<ChangeStamp>, ) -> Result<bool>

Updates the data of this state with the given value

The update is only performed if the change stamp of the state before the update matches the given expected_change_stamp. In this case, the method returns true. Otherwise, the update is not performed and the method returns false.

Note that this check is not guaranteed to work reliably in all situations. If the size of the given data exceeds the internal capacity of the state (causing a reallocation) while there is another concurrent update, it may happen that the data is updated even though the change stamp is already greater than the given one.

In order to update the data without checking the change stamp, use the set method.

Errors

Returns an error if updating fails