wnf

Struct StateName

pub struct StateName { /* private fields */ }

A state name

A state name is usually represented by its “opaque value”, which is a 64-bit integer. This opaque value can be converted to a “transparent” value by XOR’ing with the magic number 0x41C64E6DA3BC0074. This transparent value encodes certain properties of the state name in its bits. The set of these properties is represented by the StateNameDescriptor type. Use the provided TryFrom/TryInto implementations to convert between a StateName (represented by its opaque value) and the corresponding StateNameDescriptor.

Implementations

impl StateName

pub const fn from_opaque_value(opaque_value: u64) -> Self

Creates a StateName from the given opaque value

pub const fn opaque_value(self) -> u64

Returns the opaque value of this StateName