wnf

Trait AnyBitPattern

pub unsafe trait AnyBitPattern:
    Copy
    + Send
    + Sized
    + 'static { }

A marker trait for types for which any bit pattern is valid

This is modelled after the AnyBitPattern trait of the bytemuck crate.

In order for reading a value of a type T from a WNF state to be sound, T is required to implement AnyBitPattern (or at least CheckedBitPattern, which is implied by AnyBitPattern).

Implementation

This trait is already implemented by the wnf crate for many primitive types and types from the standard library. There are several ways to implement it for your own types:

Implement it directly, requiring unsafe code

Derive the AnyBitPattern trait of the bytemuck crate and derive this trait from it via the derive_from_bytemuck_v1 macro:

use wnf::{derive_from_bytemuck_v1, OwnedState};

#[derive(bytemuck::AnyBitPattern, bytemuck::NoUninit, Copy, Clone)]
#[repr(transparent)]
struct MyType(u32);

derive_from_bytemuck_v1!(AnyBitPattern for MyType);
derive_from_bytemuck_v1!(NoUninit for MyType);

let state: OwnedState<MyType> = OwnedState::create_temporary()?;
state.set(&MyType(42))?;
let data = state.get()?;

assert_eq!(data.0, 42);

Derive the IntoBytes trait of the zerocopy crate and derive this trait from it via the derive_from_zerocopy macro:

use wnf::{derive_from_zerocopy, OwnedState};

#[derive(zerocopy_derive::IntoBytes, zerocopy_derive::FromBytes, Copy, Clone)]
#[repr(transparent)]
struct MyType(u32);

derive_from_zerocopy!(AnyBitPattern for MyType);
derive_from_zerocopy!(NoUninit for MyType);

let state: OwnedState<MyType> = OwnedState::create_temporary()?;
state.set(&MyType(42))?;
let data = state.get()?;

assert_eq!(data.0, 42);

Safety

Implementing this trait for a type T is sound if any initialized sequence of bytes (“bit pattern”) with the same layout (i.e. size and alignment) as T is a valid T.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.