wnf

Trait Read

pub trait Read<D>:
    Sealed
    + Send
    + 'static { }

A trait for types that can be read from state data

A type T implements Read<D> if the data of a state of type T (i.e., an OwnedState<T> or a BorrowedState<'_, T>) can be read as an instance of type D, where D is either T or Box<T>.

This is used to abstract over types that either implement CheckedBitPattern themselves or are of the form [T] where T implements CheckedBitPattern.

This trait is sealed and cannot be implemented outside of wnf.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.